Added #if DEVICE_IS_BLE_SERVER statements in main.c

+++ src/i2c.c
@@ -79,7 +79,7 @@ float tempConv(void)
 	float celsTemp = ((175.72 * (float)tempData / 65536) - 46.85);
 	
 	/* Logging temperature data */
-	LOG_INFO("Temperature: %d\n", (uint32_t)celsTemp);
+	LOG_INFO("Temperature: %.2f\n", celsTemp);
 
 	return celsTemp;
 }
